Support Shallow Gravy
This module will now support both 3.8.x and 2015.x.

In addition, it will now work on Puppet masters that are not a
standalone node. It will call the RBAC API on the node configured as the
classifier. Obviously, for that to work, the certname must be configured
in the RBAC certificate whitelist.

It also now properly supports RBAC users in which the title != the name.